requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asttokens==3.0.0
attrs==25.3.0
azure-ai-evaluation==1.1.0
azure-ai-inference==1.0.0b9
azure-core==1.35.0
azure-identity==1.23.0
azure-monitor-opentelemetry-exporter==1.0.0b39
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cattrs==25.1.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
cryptography==45.0.5
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
docstring_parser==0.16
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
fixedint==0.1.6
Flask==3.1.1
flask-cors==6.0.1
flask-restx==1.3.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keyboard==0.13.5
keyring==24.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistralai==0.4.2
more-itertools==10.7.0
MouseInfo==0.1.3
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
openai==1.72.0
opencv-python==4.12.0.88
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.10.18
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.8
playsound==1.2.2
prompt_toolkit==3.0.51
promptflow-core==1.18.1
promptflow-devkit==1.18.1
promptflow-tracing==1.18.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
PyAudio-0.2.11-cp311-cp311-win_amd64.whl
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydash==7.0.7
PyGetWindow==0.0.9
Pygments==2.19.2
pyjokes==0.8.3
PyJWT==2.10.1
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywikihow==0.5.7
pywin32==310
pywin32-ctypes==0.2.3
pyzmq==27.0.0
ratelim==0.1.6
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-cache==1.2.1
requests-oauthlib==2.0.0
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
speedtest-cli==2.1.3
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.1
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.9.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
uvicorn==0.35.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
wikipedia==1.4.0
zipp==3.23.0
